Update damage plugins to SampleTalent
