+++ src/module/apps/damage/plugins/sampleTalent.ts
@@ -1,0 +1,67 @@
+import * as t from "io-ts";
+import { DamageHudData, DamageHudTarget } from "../data";
+import { isTalentAvailable } from "../../../util/misc";
+
+//See ./vanguard.ts and this file for an example of how to implement a new talent
+export class SampleTalent {
+  //Plugin state
+  active: boolean = false;
+
+  //AccDiffHudPlugin requirements
+  //There is most likely a way to do this in TS. If you know, tell me so I can do it right
+  //@ts-expect-error pinkie promise we will init it
+  slug: string;
+  static category: "acc" | "diff" | "talentWindow" = "talentWindow";
+  category: "acc" | "diff" | "talentWindow" = "talentWindow";
+
+  static get schema() {
+    return {
+      active: t.boolean,
+    };
+  }
+  static get schemaCodec() {
+    return t.type(this.schema);
+  }
+  get raw() {
+    return {
+      active: this.active,
+    };
+  }
+
+  //CheckboxUI requirements
+  uiElement: "checkbox" = "checkbox";
+  //Doesn't matter as of time of writing I don't think
+  rollPrecedence = 0; // higher numbers happen earlier
+
+  get uiState(): boolean {
+    return this.active;
+  }
+  set uiState(data: boolean) {
+    this.active = data;
+
+    console.log("BEING SET, active = " + this.active);
+  }
+  // this talent is only visible when the owner has talent
+  // only enabled if conditions are satisfied
+  visible = false;
+  disabled = false;
+
+  //RollModifier requirements
+  //We do nothing to modify the roll
+  modifyRoll(roll: string): string {
+    return roll;
+  }
+
+  //Dehydrated requirements
+  hydrate(data: DamageHudData, target?: DamageHudTarget) {
+    // Check if actor has talent
+    if (!isTalentAvailable(data.lancerActor, this.slug)) return;
+
+    //Figure out whether we are in a Handshake Etiquette situation
+    this.active = this.talent(data, target);
+    this.visible = true;
+  }
+
+  //@ts-expect-error pinkie promise we will init it
+  talent(data: DamageHudData, target?: DamageHudTarget): boolean;
+}
